Bug Fix & Comments

+++ src/automobiledealer/Model/Vehicle/Truck.java
@@ -10,15 +10,15 @@ package automobiledealer.Model.Vehicle;
  * @author AkbarFauzy
  */
 public class Truck extends Vehicle{
-    private double loadCapacity;
+    private int loadCapacity;
 
-    public Truck(String _registrationNumber, String _name, String _brand, String _color, int _numWheel, double _weight, int _numDoors, String _transmission, int _price, String _fuelType, int _horsePower, String _status, double _loadCapacity) {
+    public Truck(String _registrationNumber, String _name, String _brand, String _color, int _numWheel, double _weight, int _numDoors, String _transmission, int _price, String _fuelType, int _horsePower, String _status, int _loadCapacity) {
         super(_registrationNumber, _name, _brand, _color, _numWheel, _weight, _numDoors, _transmission, _price, _fuelType, _horsePower, _status);
         this.loadCapacity = _loadCapacity;
     }
 
 
-    public void setLoadCapacity(double loadCapacity){
+    public void setLoadCapacity(int loadCapacity){
         this.loadCapacity = loadCapacity;
     }
